Make axis and button notifications work
This support is *very* specific to the Logitech Dual-Action gamepad.

To do better I would need to implement USBReadParser on the gamepad
class, pass it into GetReportDescr, parse the results, store sufficient
information to pre-populate prevState with the correct types, as well as
enough data to programmatically parse the bits out of reports and update
the correct prevState entries.